+++ lib/logstorage/filter_phrase_test.go
@@ -2,6 +2,8 @@ package logstorage
 
 import (
 	"testing"
+
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fs"
 )
 
 func TestMatchPhrase(t *testing.T) {
@@ -49,8 +51,6 @@ func TestFilterPhrase(t *testing.T) {
 	t.Parallel()
 
 	t.Run("single-row", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -124,8 +124,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("const-column", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "other-column",
@@ -235,8 +233,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("dict", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -292,8 +288,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("strings", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -364,8 +358,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("uint8", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -431,8 +423,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("uint16", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -497,8 +487,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("uint32", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -563,8 +551,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("uint64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -628,8 +614,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("int64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -693,8 +677,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("float64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -800,8 +782,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("ipv4", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -898,8 +878,6 @@ func TestFilterPhrase(t *testing.T) {
 	})
 
 	t.Run("timestamp-iso8601", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "_msg",
@@ -981,4 +959,7 @@ func TestFilterPhrase(t *testing.T) {
 		}
 		testFilterMatchForColumns(t, columns, pf, "_msg", nil)
 	})
+
+	// Remove the remaining data files for the test
+	fs.MustRemoveAll(t.Name())
 }