+++ lib/logstorage/filter_string_range_test.go
@@ -2,6 +2,8 @@ package logstorage
 
 import (
 	"testing"
+
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fs"
 )
 
 func TestMatchStringRange(t *testing.T) {
@@ -27,8 +29,6 @@ func TestFilterStringRange(t *testing.T) {
 	t.Parallel()
 
 	t.Run("const-column", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -86,8 +86,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("dict", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -143,8 +141,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("strings", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -188,8 +184,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("uint8", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -241,8 +235,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("uint16", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -294,8 +286,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("uint32", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -347,8 +337,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("uint64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -400,8 +388,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("int64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -453,8 +439,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("float64", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -506,8 +490,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("ipv4", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "foo",
@@ -567,8 +549,6 @@ func TestFilterStringRange(t *testing.T) {
 	})
 
 	t.Run("timestamp-iso8601", func(t *testing.T) {
-		t.Parallel()
-
 		columns := []column{
 			{
 				name: "_msg",
@@ -623,4 +603,7 @@ func TestFilterStringRange(t *testing.T) {
 		}
 		testFilterMatchForColumns(t, columns, fr, "_msg", nil)
 	})
+
+	// Remove the remaining data files for the test
+	fs.MustRemoveAll(t.Name())
 }